Decoupled the optimizer from Program.cs
Massively decoupled the optimizer from Program.cs, it is now its own Optimizer class.

Attempted to split the "display code" from the "optimize code" as much as possible.

ChangePriority() now correctly catches the Win32 exception caused by protected processes which don't allow you to view their priority or do but don't allow you to change it. ChangePriority() will emit an error to screen if ShowErrorCodes is TRUE

Restore() correctly checks whether explorer.exe was killed before starting a new process, the previous implementation would start explorer regardless of whether it was killed or not. This is thanks to the FlagsUsedForOptimize member field which stores which flags were used on the last call to Optimize().