Corrected wording and removed hard coded Thread.Sleep(5000)
The wording for when restore completes has been changed so that it no longer says "restored to normal priority" and now just says "restored"; this is because each process now independently restores to its original priority which may not be "normal priority".

Also removed the hard coded literal 5000 in the Thread.Sleep() in Command_OptimizeWithFlags, it now correctly uses _optimizeWaitTimeMillis instead.