Fully implemented ChangeAffinity
Improved documentation and improved readability of couple of variable names etc

The primary feature of this commit is the ChangeAffinity method which enables you to modify the affinity of a process, you can provide the new ProcessAffinity bitmask enum to specify which cores should be enabled/disabled. Set the OptimizeAffinity flag on OptimizeFlags to enable this functionality for systems with 4 or more cores.

The optimizer now remembers which processes have been modified with the _changedProcesses list. Whenever a process is modified(priority or affinity changed), this change will be committed to the _changedProcesses list; the Restore method will now go through and reverse any of the modifications made during optimization. This improves performance as the previous implementation would get the process list again and loop through it, with this implementation only explicitly modified processes will be looped over.