Whitelist is now a Resource Asset
The default auto-generated whitelist file is now an embedded asset rather than being constructed in a method as was before. This allows for a more robust way of updating the default whitelist as you can just edit it through in the visual studio text edtior and can fit way more items in it.

Also added a generic `OpenWithDefaultProgram` method which avoids boilerplate code when wanting to open a file with its default windows program. Used in the help menu for loading the README and LICENCE files and in the options menu for editing the whitelist file.