Added new "IsGame" concept.
The functionality of the optimizer is changing, previously we were limited to a binary optimization technique, either you were a whitelisted process or you weren't, this commit introduces a third identifier, IsGame. The optimizer can use this new identifier to do more complex optimization, particularly with Affinity, i.e putting both whitelisted processes on the last few cores whilst leaving games on the first few, however not affecting the priority of the whitelisted processes.