Update Optimizer.cs
Improved code docs.

Fully implemented compatibility with IWhitelistedProcessIdentifierSource and IGameProcessIdentifierSource.

Refreshing the list of Whitelisted and Game processes is now the responsibility of the optimizer, the identifiers are now private to reflect this (they cannot be externally refreshed).

Games are now considered "Whitelisted" by default (this fixes a bug where the affinity optimizer would correctly put games on the first 4 cores but then the other code would reduce the game(s) to Idle priority because it didn't know it was Whitelisted.