Fixed massive oversight in BitMask.
The BitMask helper class ModifyBitRange method was broken, it did not allow users to set the 64th bit of a 64 bit number due to the way it was counting the number of bits in a nint. Essentially We needed to be using nuint.MaxValue rather than nint.MaxValue because nint has one less binary digit than the uint variant.

I've also added a testing project and included a plethora of tests for the Bitmask class to ensure proper working functionality.