Update ProcessorLayoutInformation.cs
Added a cache for the GetCurrentProcessorLayout method, this method is expensive and should only need to be called once per session, so we cache the result internally and pass it back when another request is made.

The "coresPerCCX" param for the default `ProcessorLayout` is now `0`, meaning the layout will now correctly report `HasCoreComplexes` as false.