Create IProcess.cs
The `IProcess` interface provides an interface to all the properties of a Process object that the `Optimizer` requires, this is so that we may mock the `Process` in our tests. The `ProcessWrapper` is a simple wrapper object around a `Process` so that we may pass a `Process` as an `IProcess` (because .NET doesn't support casting an object to an interface with the same public methods unless that object explicitly implements the interface).